lab3-rest-api-flask-sqlite/requirements.txt
Flask
PyJWT
requests

